Reorganization of tests
- common code moved into `common` dir
