+++ test/src/CSVOutput_test.cpp
@@ -1,4 +1,5 @@
 #include "gmock/gmock.h"
+#include "GodzillaApp_test.h"
 #include "FENonlinearProblem_test.h"
 #include "godzilla/Problem.h"
 #include "godzilla/CSVOutput.h"
