+++ test/src/FENonlinearProblem_test.cpp
@@ -1,11 +1,11 @@
 #include "gmock/gmock.h"
+#include "GodzillaApp_test.h"
 #include "FENonlinearProblem_test.h"
-#include "godzilla/CallStack.h"
+#include "godzilla/MeshObject.h"
+#include "godzilla/InitialCondition.h"
 #include "godzilla/Factory.h"
-#include "GTestFENonlinearProblem.h"
 #include "godzilla/Parameters.h"
 #include "godzilla/LineMesh.h"
-#include "godzilla/UnstructuredMesh.h"
 #include "godzilla/InitialCondition.h"
 #include "godzilla/ConstantInitialCondition.h"
 #include "godzilla/DirichletBC.h"
